Change best buy
Added a check for low payback period to trigger buying 10 buildings at a
time.  This is a compromise between speed and efficiency and may be too
slow for some situations.
Similarly, if the pp < 1 for upgrades, just buy it immediately
Added a check for dragon level to limit purchasing beyond what's
required for sacrificing.  Again this slows down purchasing for more
efficient buys and may end up being too slow.