Quick fix for Stock Market overspending
Closes #64 (hopefully)
Disables the stock market for 1 hour to hopefully fix (or at least delay) the bot from spending all of the cookies on the stock market so other things can be done. 
Stock market still needs to have saving implemented.
Adds a variable which can be used to delay things after reincarnation.